+++ src/api/projects/members/members.rs
@@ -13,6 +13,7 @@ use crate::api::endpoint_prelude::*;
 
 /// Query a members of a project.
 #[derive(Debug, Builder)]
+#[builder(setter(strip_option))]
 pub struct ProjectMembers<'a> {
     /// The project to query for membership.
     #[builder(setter(into))]