+++ CHANGELOG.md
@@ -27,6 +27,8 @@ been added.
   * The `api::projects::repository::CreateFile` endpoint now properly upgrades
     the encoding when attempting to encode binary contents using
     `Encoding::Text`.
+  * The `api::projects::CreateProject` and `api::projects::EditProject`
+    endpoints now accepts plain strings in its `import_url` field.
 
 ## Changes